sympy.py Small overhaul of sympy_solver to handle summations differently and it's internal logic. New helper functions and judgement based on parity. Rejigged output to only include solutions, and binary factorisation at the top of the file.
Add new experiments to the top of PrimeFactorizationEquationGenerator

Update ReHandler.py to have new helper functions

verification.py
Perhaps not the best module, but it is designed to help check the answers to the system of equations.
At the moment it only prints out the binary factorisation, but should eventually made to check automatically.

 Changes to be committed:
	new file:   17x17semiprimesFirst100.txt
	modified:   PrimeFactorizationEquationGenerator.py
	modified:   ReHandler.py
	modified:   sympy_solver.py
	new file:   verification.py