Change equation checking
Check to see that equations are not False, or raise a
ContradictionException. Really this should live somewhere else, as it
depends on the EquationSolver infrastructure, not just sympy, but that's
for another day.
Also add expressions_to_variables helper, which unions all of the atoms
of a list of equations.