Update sequential solver
Add option to sort variables by an arbitrary comparison function.
Add sub_next method, which brings a given variable to the front of the
substitution queue.
Return the variables that were substituted when sub_var is done.
If an equation has one variable left, substitute that one in next.
Fix bugs and update doctests.