Initial push of sympy_subs
A library of functions that should be used to carry out substitutions,
either singularly or as a group. This is partly so that we can
centralise this important functionality and streamline it in future.
For now point to the standard sympy function, while we develop something
nicer than string mangling and faster than sympy